+++ website/client/src/components/challenges/sidebar.vue
@@ -89,70 +89,14 @@
 import throttle from 'lodash/throttle';
 import FilterSidebar from '@/components/ui/filterSidebar';
 import FilterGroup from '@/components/ui/filterGroup';
+import CategoryOptions from '@/../../common/script/content/categoryOptions';
 
 export default {
   components: { FilterGroup, FilterSidebar },
   data () {
     return {
       categoryFilters: [],
-      categoryOptions: [
-        {
-          label: 'habitica_official',
-          key: 'habitica_official',
-        },
-        {
-          label: 'academics',
-          key: 'academics',
-        },
-        {
-          label: 'advocacy_causes',
-          key: 'advocacy_causes',
-        },
-        {
-          label: 'creativity',
-          key: 'creativity',
-        },
-        {
-          label: 'entertainment',
-          key: 'entertainment',
-        },
-        {
-          label: 'finance',
-          key: 'finance',
-        },
-        {
-          label: 'health_fitness',
-          key: 'health_fitness',
-        },
-        {
-          label: 'hobbies_occupations',
-          key: 'hobbies_occupations',
-        },
-        {
-          label: 'location_based',
-          key: 'location_based',
-        },
-        {
-          label: 'mental_health',
-          key: 'mental_health',
-        },
-        {
-          label: 'getting_organized',
-          key: 'getting_organized',
-        },
-        {
-          label: 'self_improvement',
-          key: 'self_improvement',
-        },
-        {
-          label: 'spirituality',
-          key: 'spirituality',
-        },
-        {
-          label: 'time_management',
-          key: 'time_management',
-        },
-      ],
+      categoryOptions: CategoryOptions,
       membershipFilters: [],
       membershipOptions: [
         {
