+++ website/client/src/components/groups/groupFormModal.vue
@@ -379,6 +379,7 @@ import informationIcon from '@/assets/svg/information.svg';
 
 import { MAX_SUMMARY_SIZE_FOR_GUILDS } from '@/../../common/script/constants';
 import { userStateMixin } from '../../mixins/userState';
+import CategoryOptions from '@/../../common/script/content/categoryOptions';
 
 // @TODO: Not sure the best way to pass party creating status
 // Since we need the modal in the header, passing props doesn't work
@@ -410,64 +411,7 @@ export default {
         allowGuildInvitationsFromNonMembers: true,
         bannedWordsAllowed: null,
       },
-      categoryOptions: [
-        {
-          label: 'habitica_official',
-          key: 'habitica_official',
-        },
-        {
-          label: 'academics',
-          key: 'academics',
-        },
-        {
-          label: 'advocacy_causes',
-          key: 'advocacy_causes',
-        },
-        {
-          label: 'creativity',
-          key: 'creativity',
-        },
-        {
-          label: 'entertainment',
-          key: 'entertainment',
-        },
-        {
-          label: 'finance',
-          key: 'finance',
-        },
-        {
-          label: 'health_fitness',
-          key: 'health_fitness',
-        },
-        {
-          label: 'hobbies_occupations',
-          key: 'hobbies_occupations',
-        },
-        {
-          label: 'location_based',
-          key: 'location_based',
-        },
-        {
-          label: 'mental_health',
-          key: 'mental_health',
-        },
-        {
-          label: 'getting_organized',
-          key: 'getting_organized',
-        },
-        {
-          label: 'recovery_support_groups',
-          key: 'recovery_support_groups',
-        },
-        {
-          label: 'spirituality',
-          key: 'spirituality',
-        },
-        {
-          label: 'time_management',
-          key: 'time_management',
-        },
-      ],
+      categoryOptions: CategoryOptions,
       showCategorySelect: false,
       members: [],
       inviteMembers: false,
