+++ website/common/script/content/categoryOptions.js
@@ -1,0 +1,60 @@
+const categoryOptions = [
+  {
+    label: 'habitica_official',
+    key: 'habitica_official',
+  },
+  {
+    label: 'academics',
+    key: 'academics',
+  },
+  {
+    label: 'advocacy_causes',
+    key: 'advocacy_causes',
+  },
+  {
+    label: 'creativity',
+    key: 'creativity',
+  },
+  {
+    label: 'entertainment',
+    key: 'entertainment',
+  },
+  {
+    label: 'finance',
+    key: 'finance',
+  },
+  {
+    label: 'health_fitness',
+    key: 'health_fitness',
+  },
+  {
+    label: 'hobbies_occupations',
+    key: 'hobbies_occupations',
+  },
+  {
+    label: 'location_based',
+    key: 'location_based',
+  },
+  {
+    label: 'mental_health',
+    key: 'mental_health',
+  },
+  {
+    label: 'getting_organized',
+    key: 'getting_organized',
+  },
+  {
+    label: 'self_improvement',
+    key: 'self_improvement',
+  },
+  {
+    label: 'spirituality',
+    key: 'spirituality',
+  },
+  {
+    label: 'time_management',
+    key: 'time_management',
+  },
+];
+
+export default categoryOptions;
